requirements.txt
fastapi==0.70.0
psycopg2-binary==2.9.1
uvicorn==0.15.0
sqlalchemy
python-dotenv